twitter-producer/requirements.txt
kafka-python
oauthlib==1.0.3
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
tweepy
datetime
scikit-learn==0.24.1
joblib==1.0.1
six==1.10.0
threadpoolctl==2.1.0
